requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
frozenlist==1.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.27.8
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.16
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==2.0.3
Werkzeug==2.3.6
yarl==1.9.2
